Refactor krun. Add tests for the scheduler and time estimate formatter. These show that a --reboot makes progress through the schedule, and should help prevent an infinite reboot loop.
